// Neo4Secure/Neo4Secure/src/app/admin/modules/users/interfaces/user.interface.ts
// src/app/admin/modules/users/interfaces/user.interface.ts
export interface IUsuario {
  id_usuario?: string;
  nombre?: string;
  apellido?: string;
  email?: string;
  pais?: string;
  ciudad?: string;
  telefono?: string;
  fecha_nacimiento?: string;
  fecha_registro?: string;
  nit?: string;
}